fixing timer job lock issue if fetch more than 1 record
